requirements.txt
asgiref==3.8.1
Django==5.1.5
django-allauth==65.3.1
django-tinymce==4.1.0
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
